+++ src/prepdir/__init__.py
@@ -1,6 +1,6 @@
 """
 prepdir - Directory traversal utility to prepare project contents for review
 """
-from .main import run
+from .main import run, validate_output_file
 
 __version__ = "0.13.0"